+++ Content.Client/_CD/Engraving/EngraveableSystem.cs
@@ -1,0 +1,5 @@
+using Content.Shared._CD.Engraving;
+
+namespace Content.Client._CD.Engraving;
+
+public sealed class EngraveableSystem : SharedEngraveableSystem;
