+++ Content.Server/_CD/Engraving/EngraveableSystem.cs
@@ -1,50 +1,34 @@
 using Content.Server.Administration;
 using Content.Server.Administration.Logs;
-using Content.Server.Popups;
+using Content.Shared._CD.Engraving;
 using Content.Shared.Database;
 using Content.Shared.Popups;
-using Content.Shared.Examine;
 using Content.Shared.Verbs;
 using Robust.Shared.Player;
-using Robust.Shared.Utility;
 
 namespace Content.Server._CD.Engraving;
 
-public sealed class EngraveableSystem : EntitySystem
+public sealed class EngraveableSystem : SharedEngraveableSystem
 {
     [Dependency] private readonly IAdminLogManager _adminLogger = default!;
-    [Dependency] private readonly PopupSystem _popup = default!;
+    [Dependency] private readonly SharedPopupSystem _popup = default!;
     [Dependency] private readonly QuickDialogSystem _dialog = default!;
 
     public override void Initialize()
     {
         base.Initialize();
 
-        SubscribeLocalEvent<EngraveableComponent, ExaminedEvent>(OnExamined);
-        SubscribeLocalEvent<EngraveableComponent, GetVerbsEvent<ActivationVerb>>(AddEngraveVerb);
+        SubscribeLocalEvent<Shared._CD.Engraving.EngraveableComponent, GetVerbsEvent<ActivationVerb>>(AddEngraveVerb);
     }
 
-    private void OnExamined(Entity<EngraveableComponent> ent, ref ExaminedEvent args)
-    {
-        var msg = new FormattedMessage();
-        msg.AddMarkupOrThrow(Loc.GetString(ent.Comp.EngravedMessage == string.Empty
-            ? ent.Comp.NoEngravingText
-            : ent.Comp.HasEngravingText));
-
-        if (ent.Comp.EngravedMessage != string.Empty)
-            msg.AddMarkupOrThrow(Loc.GetString(ent.Comp.EngravedMessage));
-
-        args.PushMessage(msg, 1);
-    }
-
-    private void AddEngraveVerb(Entity<EngraveableComponent> ent, ref GetVerbsEvent<ActivationVerb> args)
+    private void AddEngraveVerb(Entity<Shared._CD.Engraving.EngraveableComponent> ent, ref GetVerbsEvent<ActivationVerb> args)
     {
         // First check if it's already been engraved. If it has, don't let them do it again.
         if (ent.Comp.EngravedMessage != string.Empty)
             return;
 
         // We need an actor to give the verb.
-        if (!EntityManager.TryGetComponent(args.User, out ActorComponent? actor))
+        if (!TryComp<ActorComponent>(args.User, out var actor))
             return;
 
         // Make sure ghosts can't engrave stuff.
@@ -62,10 +46,11 @@ public sealed class EngraveableSystem : EntitySystem
                     (string message) =>
                     {
                         // If either the actor or comp have magically vanished
-                        if (actor.PlayerSession.AttachedEntity == null || !HasComp<EngraveableComponent>(ent))
+                        if (actor.PlayerSession.AttachedEntity == null || !HasComp<Shared._CD.Engraving.EngraveableComponent>(ent))
                             return;
 
                         ent.Comp.EngravedMessage = message;
+                        Dirty(ent);
                         _popup.PopupEntity(Loc.GetString(ent.Comp.EngraveSuccessMessage),
                             actor.PlayerSession.AttachedEntity.Value,
                             actor.PlayerSession,
